ADD-MILESTONE: wraps for normal estimation + segmentation

+++ src/diffCheck/segmentation/DFSegmentation.cc
@@ -12,7 +12,7 @@ namespace diffCheck::segmentation
         int minClusterSize,
         bool useKnnNeighborhood,
         int knnNeighborhoodSize,
-        int radiusNeighborhoodSize,
+        float radiusNeighborhoodSize,
         bool colorClusters)
     {
         if (!pointCloud->HasNormals())
